Supply-Chain-Demand-Forecasting-main/requirements.txt
pandas
numpy
torch
pytorch-lightning
pytorch-forecasting
fastapi
uvicorn
shap
scikit-learn
matplotlib
google-generativeai